+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.16.0] - 2022-11-07
+### Fixed
+- Recalculate all dashboards orders in group when one dashboard changes position
+### Added
+- order field in `/dashs/id=<group_id>` endpoint for dashboard when group id is provided
+
 ## [1.15.1] - 2022-10-14
 ### Fixed
  - Wrong earliest and latest parsing